PDF_Query_Using_AstraDB/requirements.txt
streamlit 
langchain
langchain-groq
langchain-huggingface
python-dotenv
langchain-astradb
datasets
langchain_community
langchain_core
PyPDF2
sentence-transformers
pypdf